+++ app/src/main/java/com/demo/carspends/presentation/fragments/notesList/NotesListViewModel.kt
@@ -264,7 +264,7 @@ class NotesListViewModel @Inject constructor(
                 val mileage = abs(notes.first().mileage - notes.last().mileage)
                 val fuel = abs(notes.sumOf { it.liters } - notes.last().liters)
                 if (fuel <= 0 || mileage <= 0) 0.0
-                else fuel / (mileage / 100)
+                else fuel / (mileage.toDouble() / 100)
             } else 0.0
 
             _carItem = itCar.copy(
